requirements.txt
numpy
scipy
matplotlib
pyrtlsdr
tqdm